moved Ontology service into Core; reverted StatementsViewPresenter back to from a Spring component into a normal class.
