+++ .gitignore
@@ -8,6 +8,8 @@
 *.war
 *.ear
 *.settings/
+application.properties
+ogm.properties
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
